fix: Fix "call to 'nwords' is ambiguous" error
fix bits::copyBits coredump caused by negative leafNullsSize_ (#529)
introduced nwords overload functions and this caused c++ failed with the
following error:

../../bolt/serializers/PrestoSerializer.cpp:2115:35: error: call to 'nwords' is ambiguous
 2115 |     auto* nulls = nullsHolder.get(bits::nwords(rows.size()));
      |                                   ^~~~~~~~~~~~

This commit fixes it by adding explict cast on the ambiguous callsites.

+++ bolt/common/base/benchmarks/BitUtilBenchmark.cpp
@@ -161,7 +161,7 @@ void BM_forEachBit(
     size_t numBits,
     std::optional<size_t> bitToClear = std::nullopt) {
   folly::BenchmarkSuspender suspender;
-  uint32_t size = bits::nwords(numBits);
+  uint32_t size = bits::nwords(static_cast<uint64_t>(numBits));
   std::vector<uint64_t> data(size);
   uint64_t* rawData = data.data();
   bits::fillBits(rawData, 0, numBits, true);
