+++ bolt/dwio/dwrf/reader/SelectiveFlatMapColumnReader.cpp
@@ -595,7 +595,8 @@ class SelectiveFlatMapReader : public SelectiveStructColumnReaderBase {
       selectedInMap.setValid(row, true);
     }
     int32_t totalChildValues = 0;
-    rowColumnBits_.resize(bits::nwords(rows.size() * children_.size()));
+    rowColumnBits_.resize(
+        bits::nwords(static_cast<uint64_t>(rows.size() * children_.size())));
     std::fill(rowColumnBits_.begin(), rowColumnBits_.end(), 0);
     nullsInChildValues_ = false;
     for (auto i = 0; i < children_.size(); ++i) {
